requirements.txt
python-gitlab>=4.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
